next registration(2.49.10)

+++ solhub/base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout 
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -15,6 +18,10 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+    #avoiding relogging in of an already logged in user
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -24,8 +31,20 @@ def loginPage(request):
         except:
             messages.error(request, 'User does not exist')
 
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user )
+            return redirect('home')
+        else:
+            messages.error(request, 'Incorrect username or password')
+
     context = {}
     return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -48,6 +67,7 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm
     
@@ -60,9 +80,13 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk ):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not the owner of the room !')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -73,10 +97,16 @@ def updateRoom(request, pk ):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not the owner of the room !')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
+    
     
     return render (request, 'base/delete.html', {'obj':room})
